pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "archivist"
version = "0.1.0"
description = "Filesystem-based document archive for research papers and project docs"
requires-python = ">=3.12"
dependencies = [
    "click>=8.0",
    "pymupdf>=1.24",
    "jinja2>=3.0",
    "pyyaml>=6.0",
    "flask>=3.0",
    "markdown>=3.5",
]

[project.scripts]
archivist = "archivist.cli:cli"

[tool.setuptools.packages.find]
where = ["src"]
